+++ api/src/main/java/com/discoodle/api/repository/GroupsRepository.java
@@ -35,7 +35,7 @@ public interface GroupsRepository extends JpaRepository<Groups, Integer> {
     @Modifying
     @Query(value = "insert into link_groups_to_user (user_id, groups_id) VALUES (:user_id,:groups_id)", nativeQuery = true)
     @Transactional
-    void addNewMemberInGroup(@Param("user_id") Integer user_ID, @Param("groups_id") Integer groups_ID);
+    void addNewMemberInGroup(@Param("user_id") Long user_ID, @Param("groups_id") Integer groups_ID);
 
     @Modifying
     @Query(value = "insert into link_groups_to_group (groups_id, son_id) VALUES (:groups_id,:son_id)", nativeQuery = true)
