+++ api/src/main/java/com/discoodle/api/request/GroupsRequest.java
@@ -12,7 +12,7 @@ import lombok.ToString;
 @ToString
 public class GroupsRequest {
     private final Integer parent_id;
-    private final Integer user_id;
+    private final Long user_id;
     private final Integer depth;
     private final String name;
     private final String description;
